orders/src/routes/delete.router.ts
import express, { Request, Response, NextFunction } from "express";
import {
  NotFoundError,
  OrderStatus,
  requireAuth,
  NotAuthError,
} from "@magmer/common";

import { Order } from "../models/order.model";

import { OrderCancelledPublisher } from "../events/publishers/order-cancelled-publisher";
import { natsWrapper } from "../nats-wrapper";

const router = express.Router();

router.delete(
  "/api/orders/:orderId",
  requireAuth,
  async (req: Request, res: Response, next: NextFunction) => {
    const { orderId } = req.params;
    const order = await Order.findById(orderId).populate("product");

    if (!order) return next(new NotFoundError());

    if (order.userId !== req.currentUser!.id) {
      return next(new NotAuthError());
    }

    order.status = OrderStatus.Cancelled;
    await order.save();

    new OrderCancelledPublisher(natsWrapper.client).publish({
      id: order.id,
      version: order.version,
      product: {
        id: order.product.id,
      },
    });

    res.status(204).send(order);
  }
);

export { router as deleteOrdersRouter };
